+++ PresentationModel/PresentationModel.cs
@@ -44,7 +44,7 @@ namespace TP.ConcurrentProgramming.Presentation.Model
       return eventObservable.Subscribe(x => observer.OnNext(x.EventArgs.Ball), ex => observer.OnError(ex), () => observer.OnCompleted());
     }
 
-    public override void Start(int numberOfBalls, int width, int height, int borderSize)
+    public override void Start(int numberOfBalls, double width, double height, double borderSize)
     {
       layerBellow.Start(numberOfBalls, (pos, ball) =>
       {
